+++ src/rms/tuning.page.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component } from '@angular/core';
+import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
 
 import { Observable, Subject } from 'rxjs';
 
@@ -6,7 +6,6 @@ import { ControlUnit } from '../carrera';
 import { CONTROL_UNIT_PROVIDER, Driver, Settings } from '../core';
 
 // TODO: store with CU or settings?
-// TODO: initial values, mark as touched, etc.
 const MODELS = [0, 1, 2, 3, 4, 5].map(id => ({
   id: id,
   speed: null,
@@ -18,7 +17,7 @@ const MODELS = [0, 1, 2, 3, 4, 5].map(id => ({
   providers: [CONTROL_UNIT_PROVIDER],
   templateUrl: 'tuning.page.html',
 })
-export class TuningPage {
+export class TuningPage implements OnDestroy, OnInit {
 
   drivers: Observable<Driver[]>;
 
@@ -36,24 +35,28 @@ export class TuningPage {
 
   constructor(private cu: ControlUnit, private ref: ChangeDetectorRef, settings: Settings) {
     this.drivers = settings.getDrivers();
+  }
 
-    // TODO: is this kind of single debounce correct for multiple individual ranges/types?
-    // TODO: move to ngOnInit(), unsibscribe in destroy?
+  ngOnInit() {
     this.subject.debounceTime(400).subscribe((event) => {
       for (let model of (event.id !== undefined ? [this.models[event.id]] : this.models)) {
         switch (event.type) {
         case 'speed':
-          cu.setSpeed(model.id, model.speed);
+          this.cu.setSpeed(model.id, model.speed);
           break;
         case 'brake':
-          cu.setBrake(model.id, model.brake);
+          this.cu.setBrake(model.id, model.brake);
           break;
         case 'fuel':
-          cu.setFuel(model.id, model.fuel);
+          this.cu.setFuel(model.id, model.fuel);
           break;
         }
       }
     });
+  }
+
+  ngOnDestroy() {
+    this.subject.complete();
   }
 
   update(type: string, value: number, id?: number) {